requirements.txt
urllib3
requests
protobuf~=3.19.4
grpcio~=1.47.0
grpcio-tools==1.44.0

pyyaml
openpyxl
pandas
scikit-learn
jsonlines
blobfile
lz4
pyzstd
jsonpath
tabulate
datasets

locust
